Add debugging section to README
The purpose of the debugging section is to create centralized
location where anyone can contribute workarounds around issues
they were having while going through the tutorial.

The first subsection in the debussion section is "Access Denied".
This is a result of IAM misconfiguration when the access keys
being used to do not administrator access.